Move otel and config code to their own modules
