+++ test/modules/obsagent-client/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod config;
+pub mod otel_client;
+pub mod prometheus_server;
